dicectf-finals-2024/mental-poker/requirements.txt
aiohttp
fastecdsa
pycryptodome
requests
